requirements.txt
# Base app (use requirements-cpu.txt or requirements-gpu.txt for full stack)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.0.0
prometheus_client>=0.19.0
numpy>=1.24.0
